Add out of bounds check for encoding/decoding (#42)
* Fixed errors

* Add safe encoding/decoding mode with buffer overflow control

---------

+++ include/protopuf/byte.h
@@ -35,7 +35,7 @@ namespace pp {
     /// Returns the byte-distance between `begin(a)` and `begin(b)`.
     inline constexpr std::size_t begin_diff(bytes a, bytes b) {
         // `std::to_address` is used here for MSVC, ref to https://github.com/microsoft/STL/issues/1435
-        return std::to_address(a.begin()) - std::to_address(b.begin());
+        return static_cast<std::size_t>(std::to_address(a.begin()) - std::to_address(b.begin()));
     }
 }
